Task03 Габидуллина

+++ task03/src/com/example/task03/Seconds.java
@@ -23,6 +23,6 @@ public class Seconds implements TimeUnit {
 
     @Override
     public long toMinutes() {
-        return Math.round(amount / 60);
+        return Math.round((double)amount / 60);
     }
 }
